src/processing/pyproject.toml
[project]
name = "processing"
version = "0.1.0"
description = "Data processing for crop classification pipeline"
requires-python = ">=3.10"
dependencies = [
    "boto3",
    "botocore[crt]>=1.42.89",
    "geopandas",
    "odc-stac",
    "psutil",
    "pystac-client",
    "pyarrow",
    "rasterio",
    "rioxarray",
    "s3fs",
    "scipy",
    "shapely",
    "tqdm",
    "xarray",
    "zarr",
    "bottleneck",
    "dask",
    "matplotlib>=3.10.8",
]

[tool.hatch.build.targets.wheel]
packages = ["processing"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
